29-07-2018

+++ resources/views/datatable/_action.blade.php
@@ -1,4 +1,4 @@
-{!! Form::model($model, ['url' => $form_url, 'method' => 'delete', 'class' => 'form-inline'] ) !!}
-<a href="{{ $edit_url }}">Ubah</a> |
+{!! Form::model($model, ['url' => $form_url, 'method' => 'delete', 'class' => 'form-inline js-confirm', 'data-confirm' => $confirm_message] ) !!}
+<a href="{{ $edit_url }}" class="btn btn-warning">Ubah</a> |
 {!! Form::submit('Hapus', ['class'=>'btn btn-xs btn-danger']) !!}
 {!! Form::close()!!}
